vahansetu/requirements.txt
Flask
flask-cors
flask-login
werkzeug
razorpay
PyJWT
gunicorn
python-dotenv
requests
python-jose
